Placate a warning.

+++ test.c
@@ -7,8 +7,8 @@ int main()
     void *c = hm_new();
     hm_insert(c, 123, "abc");
 
-    printf("123 = %016lx\n", hm_get(c, 123));
-    printf("124 = %016lx\n", hm_get(c, 124));
+    printf("123 = %016lx\n", (uintptr_t)hm_get(c, 123));
+    printf("124 = %016lx\n", (uintptr_t)hm_get(c, 124));
 
     hm_delete(c);
     return 0;
